Bug 973447 - [settings] refactor Language panel with AMD pattern
