+++ apps/settings/elements/phone_lock.html
@@ -51,7 +51,7 @@
       </ul>
     </div>
 
-    <script src="js/phone_lock.js"></script>
+    <panel data-path="panels/phone_lock/panel"></panel>
 
   </template>
 </element>
