+++ apps/settings/js/modules/settings_service.js
@@ -91,6 +91,10 @@ define(function(require) {
                   _currentPanelId = panelId;
                   _currentPanel = panel;
 
+                  // XXX we need to remove this line in the future
+                  // to make sure we won't manipulate Settings
+                  // directly
+                  Settings._currentPanel = '#' + panelId;
                   callback();
               });
             });
